* Added: [Install] Support for Solaris' native install tool. It have nothing todo with POSIX install. * Modified: [Config] Looks like some versions of bzlib had the right prototypes in the header but the library contained the pre 1.0 and 1.0 function names. To avoid detecting it as pre 1 now I test using C++ where function prototypes are mandatory. Seen on Solaris with bzlib 1.0.1. * Added: [Compilation] Support for Solaris 9 implementation of isspace when compiling with gcc 3.x. It uses the char value as index of an array. * Fixed: [Solaris] stackdbg.c didn't request alloca. * Fixed: [MSVC][Compilation] Missing () in comparissons like: overwrite != ((un.Flags & undoOverWrite)!=0) ? True : False. Reported by Grzegorz. * Modified: [MSVC][Compilation] The command to compile the assembler files. Not yet tested. * Modified: [MSVC][Compilation] Command line option to compile the code. It looks like new versions of MSVC can't compile code that includes iostream using /GX- * Fixed: [MSVC][Compilation] An empty include path made the unified makefile useless. Reported by Grzegorz.

+++ setedit/calcu/parser.c
@@ -106,7 +106,8 @@ static char *My_strndup (char *ptr, int len)
 
 static void skipspace (void)
 {
-   while (isspace (*str))
+   // The cast is a workaround for a bug in Solaris 9 for SPARC
+   while (isspace ((int)*str))
       str ++;
 }
